JS and CSS Clock
Achieved through ‘new Date()’ in javascript and  ‘transform:
rotate(90deg)’ in css. In order to remove glitch when the different
hand hits 90 450 degrees (360 + the initial 90 degree transform), I had
to remove the transition at 90 degrees and add back, as seen in the if
else statements following the second, minute, and hour computations.